Delete fakes and add DAO support

+++ src/main/java/twitter/service/privilage/PrivilegeService.java
@@ -29,7 +29,7 @@ public interface PrivilegeService {
    * Removes privilege from app
    * @param id - privilege id
    */
-  void removePrivilege(Integer id);
+  void removePrivilege(Long id);
 
   Privilege findByName(String name);
 }
